process generics separately
